refactor(view): increase cohesion of view code
